[CL-2753] Dry run by default

+++ back/engines/commercial/multi_tenancy/lib/tasks/core/clean_up_image_records.rake
@@ -2,7 +2,13 @@
 
 namespace :cl2back do
   desc 'Remove image records not associated with resource or with nil value for image field'
-  task clean_up_image_records: :environment do
+  # Usage:
+  # Dry run (no changes): cl2back:clean_up_image_records
+  # Execute (destroys records!): cl2back:clean_up_image_records EXECUTE_CLEAN_UP_UNUSED_IMAGE_RECORDS=true
+  # We use an ENV var as this seems to work better with specs.
+  task clean_up_unused_image_records: :environment do
+    dry_run = true unless ENV.fetch('EXECUTE_CLEAN_UP_UNUSED_IMAGE_RECORDS') == 'true'
+
     Tenant.all.each do |tenant|
       puts "Processing images for tenant #{tenant.name}"
 
@@ -18,7 +24,7 @@ namespace :cl2back do
 
         ContentBuilder::LayoutImage.all.each do |image|
           if image_codes.exclude?(image.code)
-            image.destroy!
+            image.destroy! unless dry_run
             puts " destroyed layout_image with unused code field: #{image.id}"
           end
         end
@@ -41,7 +47,7 @@ namespace :cl2back do
           end
 
           if reference_found == false
-            image.destroy!
+            image.destroy! unless dry_run
             puts " destroyed text_image with unused text_reference: #{image.id}, imageable_id: #{image.imageable.id}"
           end
         end
